requirements.txt
django
djangorestframework
google-generativeai
pinecone
pillow
python-dotenv
django-cors-headers
gunicorn
pandas
